feat(editor): add dependency injection support in editors (#33)
* feat(editor): add dependency injection support in editors

The current state supports editor dependencies through the Column.params property. By adding `Factory.of`, with aurelia's container, we are wrapping the creation of an editor in a function that already resolves injected dependencies at the time the grid is bound. `Factory.of` returns a function that passes any number of arguments to the editor. When Slickgrid calls `currentEditor = new useEditor({})`, the function created by `Factory.of` passes the slick grid arguments to the editor's constructor after the injected dependencies, so order matters.

closes #18

* fix(editor): default to en when no locale

if i18n is not configured it will return undefined. we want to return 'en'