fix(gridOptions): use this.gridOptions in bind fn
using the first arg in the bind method looks at the parent context, which might not be the view-model with the `gridOptions`. For example in a repeat.for, the view-model properties are not available, which can be wrong if aurelia slickgrid in a repeat.for. Using `this.gridOptions` assures us we are using the options that are bound