fix(sorters): return -1 or 1 for invalid dates (#70)
* fix(sorters): return -1 or 1 for invalid dates

Returning zero every time value1 or value 2 is invalid will result in an incorrect sort because it would not sort one value higher than the other (e.g if value1 was 2018-05-24 and value2 was null, these would be sorted as equals). We want invalid dates to be grouped together in the sort and this accomplishes that.

We could add additional logic such that if both were invalid that we return 0, but this commit does the job just fine. Since the logic was the same for every sorter, with the exception of format and strict, I created a utility function that took all the parameters

closes #68
associates to PR #69

* fix(sorter): add null check on Date Sorter to improve performance