refactor(messaging): throw error for asyncCollection if response.bodyUsed
A user can read from the response and pass the grid the same response
(e.g. I have done this in Http Interceptors ), but javascript does not permit
reading from the response body twice. If the body is used, updateEditorCollection
will silently fail because there is no catch handler. We can check for this and throw our own
message with a few suggestions