feat(editor): Add auto commit edit flag to grid options (#96)
* feat(editor): Add auto commit flag to grid options

The auto commit flag will attempt to commit the current editor when the
editor is focused out. However, not all editors behave the same so extra
logic had to be included in the some editors. This means that the grid may not
be able to auto commit custom editors and it will be the
responsibility of the user building the editor.

checkboxEditor: Can be handled by the focusout event, but a click event
was added in the editor to make it behave more like a checkbox
dateEditor, sliderEditor: Prevent focus to the next editor
longTextEditor: Add focusout event to textarea, grid does not capture event
selectEditor: No focusout event so commit on close
others: handled by aurelia-slickgrid subscription to focusout

closes #82

* style(comments): fix whitespace and remove old comment